persist bindings
