+++ crates/input_manager/Cargo.toml
@@ -14,3 +14,4 @@ bevy = { workspace = true }
 bevy_console = { workspace = true }
 bevy_egui = { workspace = true }
 bimap = { workspace = true }
+serde_json = { workspace = true }
