+++ crates/scene_runner/src/update_world/pointer_events.rs
@@ -1,8 +1,10 @@
 use bevy::prelude::*;
-use common::structs::{ToolTips, TooltipSource};
-use input_manager::InputMap;
+use common::{
+    inputs::InputMap,
+    structs::{ToolTips, TooltipSource},
+};
 
-use crate::update_scene::pointer_results::{PointerTarget, PointerTargetInfo};
+use crate::update_scene::pointer_results::{IaToCommon, PointerTarget, PointerTargetInfo};
 use dcl::interface::ComponentPosition;
 use dcl_component::{proto_components::sdk::components::PbPointerEvents, SceneComponentId};
 
@@ -62,7 +64,7 @@ fn hover_text(
                         if info.show_feedback.unwrap_or(true) {
                             if let Some(text) = info.hover_text.as_ref() {
                                 let button = input_map
-                                    .get_input(info.button())
+                                    .get_input(info.button().to_common())
                                     .map(|b| serde_json::to_string(&b).unwrap())
                                     .unwrap_or_else(|| "(No binding)".to_owned());
                                 return Some((
