+++ crates/ui_core/src/focus.rs
@@ -1,11 +1,13 @@
 use bevy::prelude::*;
-use dcl_component::proto_components::sdk::components::common::InputAction;
-use input_manager::{Action, InputManager, InputPriority, InputType, SystemAction};
+use input_manager::{InputManager, InputPriority, InputType};
 
 use crate::ui_actions::{UiActionPriority, UiFocusActionSet};
 
 use super::ui_actions::UiActionSet;
-use common::sets::SceneSets;
+use common::{
+    inputs::{Action, CommonInputAction, SystemAction},
+    sets::SceneSets,
+};
 
 #[derive(Component)]
 pub struct Focus;
@@ -33,7 +35,7 @@ fn defocus(
     focus_elements: Query<(Entity, Ref<Focus>)>,
     mut input_manager: InputManager,
 ) {
-    let refocussed = input_manager.just_down(InputAction::IaPointer, InputPriority::Focus)
+    let refocussed = input_manager.just_down(CommonInputAction::IaPointer, InputPriority::Focus)
         || input_manager.just_down(SystemAction::Cancel, InputPriority::CancelFocus)
         || focus_elements.iter().any(|(_, focus)| focus.is_changed());
 
@@ -55,7 +57,7 @@ fn defocus(
                 InputPriority::CancelFocus,
             );
             input_manager.priorities().reserve(
-                InputType::Action(Action::Scene(InputAction::IaAny)),
+                InputType::Action(Action::Scene(CommonInputAction::IaAny)),
                 InputPriority::Focus,
             );
         } else {
@@ -65,7 +67,7 @@ fn defocus(
                 InputPriority::CancelFocus,
             );
             input_manager.priorities().release(
-                InputType::Action(Action::Scene(InputAction::IaAny)),
+                InputType::Action(Action::Scene(CommonInputAction::IaAny)),
                 InputPriority::Focus,
             );
         }
@@ -81,7 +83,7 @@ fn focus(
     for (entity, interaction, maybe_priority) in focused_elements.iter() {
         if interaction != &Interaction::None
             && input_manager.just_down(
-                InputAction::IaPointer,
+                CommonInputAction::IaPointer,
                 maybe_priority.copied().unwrap_or_default().0,
             )
         {
